Changes to CSS & Tailwind config.
So it looks like when you modify the "theme.css", any modifications made
are just added to the final CSS file when the theme is built, we wanted
to actually overwrite some of the CSS not just add to it. Upon searching
for the base css file that the make process uses to format the CSS, I
discovered https://github.com/sni/Thruk/blob/master/themes/base.css.
This includes a "Body" tag on line 315 which matches the Body tag in the
compiled version. A variable called "textBody" is used which sets some
default params such as text size, which is what we want to edit. After a
quick search of the default text size in SNI's repo (13px) I found that
textBody was being defined in tailwind.config.js.

So, I'm removing the CSS I previously added and modifying the font size
in tailwind.config.js so that the CSS is built with 12px rather than
13px.